refactor: Update form-property component to include realty ID in route
The form-property component has been updated to include the realty ID in the route path. This allows for retrieving the property data based on the ID and populating the form with the corresponding information. This change ensures that the correct property is displayed and edited when accessing the "new-property" route with a specific ID.