fix: Update FormResidentComponent to display error message for unimplemented route
The logic for updating a resident in the `updateResident` method of the `FormResidentComponent` has been fixed to display an error message indicating that the route is not yet implemented. This was achieved by calling the `_toast.error` method with the message "Please contact the support team. Route not implemented yet!".